Time: 16 ms (10.34%), Space: 8.7 MB (14.3%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        for(int i = 0;i<=x;i++){
+        for(long i = 0;i<=x;i++){
             if(i*i==x)return i;
             if(i*i>x)return i-1;
         }
